requirements.txt
# Job-O-Matic Dependencies
# Core Streamlit application dependencies

streamlit==1.33.0
requests==2.31.0
python-dotenv==1.0.0
pyyaml==6.0
sqlalchemy==2.0.0
python-docx==0.8.11
pandas==2.1.4
sqlmodel==0.0.14
Pillow==10.2.0

# Enhanced features for robust HTTP handling
urllib3==2.0.0
certifi==2023.0.0

# Development and testing (optional)
pytest==7.0.0
black==23.0.0

